# Run 1 iteration per iteration