@lbulej 2019-07-12
Thanks for the input! We are currently working on support for named configurations. When that lands (we expect roughly in 1-2 weeks), we would be happy to merge any settings that improve the usability of the suite with JMH into "jmh" configuration, so that you could then run all the benchmarks with this configuration when using JMH. Until then we would prefer to not change these values, because they significantly change operation times in benchmarks with internal loops.