@lbulej 2019-08-08
It took me a little longer than expected, but the support for named configurations that I was brewing in #191 is now merged.

All benchmarks with some configurable parameters explicitly define a configuration named `jmh` which is used by the JMH wrapper and which can hold JMH-specific overrides of default parameter values. Some benchmarks have parameters that are not configurable yet, but making them configurable is not difficult.

I'm closing this PR because #191 includes the change proposed here in form of a configuration override for JMH (even though it does not remove the internal loop). We would be happy to accept other changes to `jmh` configurations.